requirements.txt
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.23.2
python-multipart==0.0.6
httpx==0.25.1
pyvcf3==1.0.3
pyfaidx==0.7.2
typing-extensions==4.8.0 